# Replace parse_expr and destructive_add by MutableArithmetics